Restructured Generic.Properties into 3 files
Moved some code to Generic.Functions and Generic.GenState
Added viewGenState, to observe what is generated.
Adjusted generators to exercise more code, and to run faster.
Added a few more 'Summary' pretty printers.

pick 06d1ba181 Restructured Generic.Properties into 3 files, addin Generic.Functions and Generic.GenState
pick 7a8f24448 Cleaned things up, adjusted generators some
